[FIX] 코인 주문 현황 상세 조회 (SSE) 버그 및 차트 버그 해결

+++ frontend/app/src/main/java/com/a406/horsebit/DataModels.kt
@@ -64,7 +64,7 @@ data class TokenShow(
 
 data class Order(
     val price: Long,    // 호가 가격
-    val volume: Long,   // 주문량
+    val volume: Double,   // 주문량
     val priceRateOfChange: Double,  // 변동추이
 )
